requirements.txt
asgiref==3.4.1
attrs==21.3.0
Django==4.0
iniconfig==1.1.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
